pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "planner"
version = "0.1.0"
description = "Planner CLI"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Planner App" }]
dependencies = [
  "typer>=0.12",
  "google-api-python-client",
  "google-auth",
  "google-auth-oauthlib",
  "google-auth-httplib2",
  "tomli; python_version < '3.11'",
]

[project.optional-dependencies]
test = [
  "pytest>=7",
]

[project.scripts]
planner = "planner.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
